Don't reuse ceritificate serial numbers
I was testing a local HTTPS server using localhost.  If for some reason
you remove ~/.localhost, Firefox will reject the newly created
certificate saying that it has the same serial number from a previously
seen certificate, but it's a diffreent one. Just use the current UNIX
epoch as the serial number to avoid this.

Signed-off-by: Antonio Terceiro <terceiro@debian.org>